requirements.txt
moviepy>=1.0.3,<2.0.0
Pillow==9.5.0
mutagen==1.46.0
numpy==1.24.3
imageio==2.31.1
imageio-ffmpeg==0.4.8
tkinter
psutil
